Hoist the style parts of traversal.rs into style/.
